Fix an error with overlays
Something (flake-parts?) complains about anything other than a 2-argument function as an overlay
with Nixpkgs 24.05. This moves the Emacs-specific overlay to the `lib` output, which is unconstrained.

This also adds an example Home Manager configuration, which triggers the error about the
overlay (better than only noticing it downstream, as I originally did).